+++ sql/updates/h3--3.5.0--3.6.0.sql
@@ -17,12 +17,24 @@
 -- complain if script is sourced in psql, rather than via CREATE EXTENSION
 \echo Use "ALTER EXTENSION h3 UPDATE TO '3.6.0'" to load this file. \quit
 
+-- Hierarchical grid functions (hierarchy.c)
+CREATE OR REPLACE FUNCTION h3_to_center_child(h3index, resolution integer DEFAULT -1) RETURNS h3index
+    AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+    COMMENT ON FUNCTION h3_to_parent(h3index, resolution integer) IS
+'Returns the center child (finer) index contained by input index at given resolution';
+
+-- Miscellaneous H3 functions (miscellaneous.c)
+CREATE OR REPLACE FUNCTION h3_get_pentagon_indexes(resolution integer) RETURNS SETOF h3index
+    AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
+    COMMENT ON FUNCTION h3_get_res_0_indexes() IS
+'All the pentagon H3 indexes at the specified resolution.';
+
+-- type casts
 CREATE OR REPLACE FUNCTION h3index_to_bigint(h3index) RETURNS bigint
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 CREATE OR REPLACE FUNCTION bigint_to_h3index(bigint) RETURNS h3index
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
 
--- type casts
 CREATE CAST (h3index AS bigint) WITH FUNCTION h3index_to_bigint(h3index);
 CREATE CAST (bigint AS h3index) WITH FUNCTION bigint_to_h3index(bigint);